Fix imports/structure, foreign keys, fix tests
