+++ cidc_api/models/__init__.py
@@ -1,4 +1,4 @@
 from .models import *
 from .files import *
 from .schemas import *
-from .templates.example_models import *
+from .templates import *
